requirements.txt
kafka-python
msgpack-python